Use paging for S3 results

+++ pkg/cloud/services/s3/s3.go
@@ -99,24 +99,33 @@ func (s *Service) DeleteBucket() error {
 
 	// Delete machine pool user data files that did not get deleted
 	// yet by the lifecycle policy
-	out, err := s.S3Client.ListObjectsV2(&s3.ListObjectsV2Input{
-		Bucket: aws.String(bucketName),
-		Prefix: aws.String("machine-pool/"),
-	})
-	if err != nil {
-		return err
-	}
-	for _, obj := range out.Contents {
-		_, err := s.S3Client.DeleteObject(&s3.DeleteObjectInput{
+	for {
+		out, err := s.S3Client.ListObjectsV2(&s3.ListObjectsV2Input{
 			Bucket: aws.String(bucketName),
-			Key:    obj.Key,
+			Prefix: aws.String("machine-pool/"),
 		})
 		if err != nil {
 			return err
 		}
+
+		// Stop on last page of results
+		if len(out.Contents) == 0 {
+			break
+		}
+
+		log.Info("Deleting S3 files of machine pools", "count", len(out.Contents))
+		for _, obj := range out.Contents {
+			_, err := s.S3Client.DeleteObject(&s3.DeleteObjectInput{
+				Bucket: aws.String(bucketName),
+				Key:    obj.Key,
+			})
+			if err != nil {
+				return err
+			}
+		}
 	}
 
-	_, err = s.S3Client.DeleteBucket(&s3.DeleteBucketInput{
+	_, err := s.S3Client.DeleteBucket(&s3.DeleteBucketInput{
 		Bucket: aws.String(bucketName),
 	})
 	if err == nil {
